SquareGridGenerate.lua: Add semisphere support
This is a map which is topologically equivalent (homeomorphic) to a
sphere which is fairly easy to implement on a square grid.  It allows
one to have a reasonable planet simulation without the headaches of trying
to tile a sphere.

The way it works is, if we go “above” the top of the map, we end up on
the top of the map, but halfway over.  Ditto with the bottom of the map.

This way, each local neighborhood is the same (all tiles are squares;
all squares allow movement in eight directions), but the map works like
a sphere (we can more quickly move from one point to another east or west
near the north and south poles).

One issue with semispheres is that, if this form is used in a game,
once we cross the north or south pole, the map becomes a vertically
flipped mirror image.  One can either play in the “flipped” form, or
we have the user interface “correct” their orientation after they
cross the pole.