We now avoid longer "zig zag" paths
In order to make our pathfinder not use exponential time, we will
assume that, for a given path, the shortest path between any two
nodes (A and Z) X spaces apart with a A->B->C->...->Z path will have
A->B be the shortest one-step path between A and any node which takes
us to Z in X steps.

This works for the usual square, isomorphic square, and hex grids, but
can fail with some "adventure map" grids.  Let's take this:

	A -- 2 --> B -- 1 -\
	|                   +--> D
	+ -- 1 --> C -- 9 -/

Here, A->B->D is three points total, so is the fastest path from A to D.
That's two points to go from A to B and 1 point to go from B to D.

A->C->D is longer: 10 points.  One point to go from A to C and nine points
to go from C to D.  However, our pathfinder will choose A->C->D
because it has the fastest path from A to an adjacent node which takes us
to D in two steps.

So, it's not a perfect pathfinder for any given grid, but it works for
for "normal" wargame grids used in the context of a procedural/random
map generator.